fix: resolve compilation issues across iOS and Android
- Android: fix xor operator on Byte types in CryptoService (needs Int cast)
- Android: replace deprecated setTargetResolution with ResolutionSelector API
- Android: remove unused android.util.Size import
- iOS: add @unchecked Sendable to APIService and KeychainService
- iOS: add Sendable conformance to ShiftDay struct

+++ apps/android/app/src/main/java/org/llamenos/hotline/crypto/CryptoService.kt
@@ -494,7 +494,7 @@ class CryptoService @Inject constructor() {
         val secretBytes = ourSecret.chunked(2).map { it.toInt(16).toByte() }.toByteArray()
         val publicBytes = theirPublic.chunked(2).map { it.toInt(16).toByte() }.toByteArray()
         val sharedBytes = ByteArray(32) { i ->
-            (secretBytes.getOrElse(i) { 0 } xor publicBytes.getOrElse(i) { 0 }).toByte()
+            (secretBytes.getOrElse(i) { 0.toByte() }.toInt() xor publicBytes.getOrElse(i) { 0.toByte() }.toInt()).toByte()
         }
 
         return sharedBytes.joinToString("") { "%02x".format(it) }
